Allow ability to run Docker-in-Docker-in-OpenShift for testing
A bit of a complicated PR, but essentially, after much trial-and-error,
this is required in order to get OpenShift running within
Docker-in-Docker within a CentOS7 container.

Essentially, what happens, is that OpenShift tries to start Kubelet but
is unable to due to cgroup access within the CentOS7 Container host.

See issue: https://github.com/kubernetes/kubernetes/issues/43856

In the end, it is instead required to:

 1. Generate the default configuration for the all-in-one OpenShift
 origin container
 2. Edit node-config.yaml to ignore QoS containers
 3. Start the OpenShift origin container.

This will only occur if the `deployment` value is set to `test`

Another note is that `selinux` has been added as a tag to "Set SELinux
context for openshift shared dirs" due to selinux not being deployed /
used on CentOS7 containers. This is added so we can use --skip-tag
"selinux" if we are deciding to deploy to CentOS7 containers.