Use image name with registry and comment image tagging
Buildah prepends `image_name` with `localhost` make an image like
`bamachrn/python:release` show up as `localhost/bamachrn/python:release`
when checked using `builda images`. To workaround this, we can directly
use image name with registry in various stages.

Also we don't need to tag an image name to image name with registry
because we're building it with full registry name in the first place.